requirements.txt
Django>=5.2.4
qrcode
Pillow
requests
python-dotenv 